requirements.txt
fastapi==0.115.12
gTTS==2.5.4
joblib==1.4.2
keras==3.9.1
keras-facenet==0.3.2
mediapipe==0.10.21
mtcnn==1.0.0
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
pillow==11.1.0
pydantic==2.11.1
pydantic_core==2.33.0
pygame==2.6.1
pymongo==4.11.3
python-dotenv==1.1.0
pyttsx3==2.98
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
uvicorn==0.34.0
opencv-python-headless==4.11.0.86
aiosmtplib==3.0.2
fastapi-mail==1.4.2 
email_validator==2.2.0
dnspython==2.7.0